+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2.6.3
+
+- Fixed message parsing breaking on translation strings with indexed placeholders
+
 ## 2.6.2
 
 - Fixed format codes being replaced by regex capture groups in custom messages